+++ packages/ragbits-core/tests/unit/vector_stores/test_qdrant.py
@@ -1,10 +1,11 @@
-import typing
 from unittest.mock import AsyncMock
+from uuid import UUID
 
 import pytest
 from pydantic import ValidationError
 from qdrant_client.http import models
 
+from ragbits.core.embeddings.noop import NoopEmbedder
 from ragbits.core.vector_stores.base import VectorStoreEntry
 from ragbits.core.vector_stores.qdrant import QdrantVectorStore
 
@@ -12,17 +13,15 @@ from ragbits.core.vector_stores.qdrant import QdrantVectorStore
 @pytest.fixture
 def mock_qdrant_store() -> QdrantVectorStore:
     return QdrantVectorStore(
-        client=AsyncMock(),
-        index_name="test_collection",
+        client=AsyncMock(), index_name="test_collection", embedder=NoopEmbedder(return_values=[[[0.1, 0.2, 0.3]]])
     )
 
 
 async def test_store(mock_qdrant_store: QdrantVectorStore) -> None:
     data = [
         VectorStoreEntry(
-            id="1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8",
-            key="test_key",
-            vector=[0.1, 0.2, 0.3],
+            id=UUID("1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8"),
+            text="test_key",
             metadata={
                 "content": "test content",
                 "document_meta": {
@@ -39,23 +38,26 @@ async def test_store(mock_qdrant_store: QdrantVectorStore) -> None:
 
     mock_qdrant_store._client.collection_exists.assert_called_once()  # type: ignore
     mock_qdrant_store._client.create_collection.assert_called_once()  # type: ignore
-    mock_qdrant_store._client.upload_collection.assert_called_with(  # type: ignore
-        collection_name="test_collection",
-        vectors=[[0.1, 0.2, 0.3]],
-        payload=[
-            {
-                "document": "test_key",
-                "document_meta": {
-                    "title": "test title",
-                    "source": {"path": "/test/path"},
-                    "document_type": "test_type",
-                },
-                "content": "test content",
-            }
-        ],
-        ids=["1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8"],
-        wait=True,
-    )
+    mock_qdrant_store._client.upload_points.assert_called_once()  # type: ignore
+    call_kwargs = mock_qdrant_store._client.upload_points.call_args.kwargs  # type: ignore
+    call_points = list(call_kwargs["points"])
+
+    assert call_kwargs["collection_name"] == "test_collection"
+    assert len(call_points) == 1
+    assert call_points[0].id == "1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8"
+    assert call_points[0].vector == {"text": [0.1, 0.2, 0.3]}
+    assert call_points[0].payload == {
+        "id": UUID("1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8"),
+        "text": "test_key",
+        "metadata": {
+            "content": "test content",
+            "document_meta": {
+                "title": "test title",
+                "source": {"path": "/test/path"},
+                "document_type": "test_type",
+            },
+        },
+    }
 
 
 async def test_retrieve(mock_qdrant_store: QdrantVectorStore) -> None:
@@ -64,30 +66,36 @@ async def test_retrieve(mock_qdrant_store: QdrantVectorStore) -> None:
             models.ScoredPoint(
                 version=1,
                 id="1f908deb-bc9f-4b5a-8b73-2e72d8b44dc5",
-                vector=[0.12, 0.25, 0.29],
+                vector={"text": [0.12, 0.25, 0.29]},
                 score=0.9,
                 payload={
-                    "document": "test_key 1",
-                    "content": "test content 1",
-                    "document_meta": {
-                        "title": "test title 1",
-                        "source": {"path": "/test/path-1"},
-                        "document_type": "txt",
+                    "id": "1f908deb-bc9f-4b5a-8b73-2e72d8b44dc5",
+                    "text": "test_key 1",
+                    "metadata": {
+                        "content": "test content 1",
+                        "document_meta": {
+                            "title": "test title 1",
+                            "source": {"path": "/test/path-1"},
+                            "document_type": "txt",
+                        },
                     },
                 },
             ),
             models.ScoredPoint(
                 version=1,
                 id="827cad0b-058f-4b85-b8ed-ac741948d502",
-                vector=[0.13, 0.26, 0.30],
+                vector={"text": [0.13, 0.26, 0.30]},
                 score=0.9,
                 payload={
-                    "document": "test_key 2",
-                    "content": "test content 2",
-                    "document_meta": {
-                        "title": "test title 2",
-                        "source": {"path": "/test/path-2"},
-                        "document_type": "txt",
+                    "id": "827cad0b-058f-4b85-b8ed-ac741948d502",
+                    "text": "test_key 2",
+                    "metadata": {
+                        "content": "test content 2",
+                        "document_meta": {
+                            "title": "test title 2",
+                            "source": {"path": "/test/path-2"},
+                            "document_type": "txt",
+                        },
                     },
                 },
             ),
@@ -99,17 +107,17 @@ async def test_retrieve(mock_qdrant_store: QdrantVectorStore) -> None:
         {"content": "test content 2", "title": "test title 2", "vector": [0.13, 0.26, 0.30]},
     ]
 
-    entries = await mock_qdrant_store.retrieve([0.12, 0.25, 0.29])
+    query_results = await mock_qdrant_store.retrieve("query")
 
-    assert len(entries) == len(results)
-    for entry, result in zip(entries, results, strict=True):
-        assert entry.metadata["content"] == result["content"]
-        assert entry.metadata["document_meta"]["title"] == result["title"]
-        assert entry.vector == result["vector"]
+    assert len(query_results) == len(results)
+    for query_result, result in zip(query_results, results, strict=True):
+        assert query_result.entry.metadata["content"] == result["content"]
+        assert query_result.entry.metadata["document_meta"]["title"] == result["title"]
+        assert query_result.vectors["text"] == result["vector"]
 
 
 async def test_remove(mock_qdrant_store: QdrantVectorStore) -> None:
-    ids_to_remove = ["1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8"]
+    ids_to_remove = [UUID("1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8")]
 
     await mock_qdrant_store.remove(ids_to_remove)
 
@@ -117,7 +125,7 @@ async def test_remove(mock_qdrant_store: QdrantVectorStore) -> None:
     mock_qdrant_store._client.delete.assert_called_with(  # type: ignore
         collection_name="test_collection",
         points_selector=models.PointIdsList(
-            points=typing.cast(list[int | str], ids_to_remove),
+            points=["1c7d6b27-4ef1-537c-ad7c-676edb8bc8a8"],
         ),
     )
 
@@ -129,30 +137,36 @@ async def test_list(mock_qdrant_store: QdrantVectorStore) -> None:
             models.ScoredPoint(
                 version=1,
                 id="1f908deb-bc9f-4b5a-8b73-2e72d8b44dc5",
-                vector=[0.12, 0.25, 0.29],
+                vector={"text": [0.12, 0.25, 0.29]},
                 score=0.9,
                 payload={
-                    "document": "test_key 1",
-                    "content": "test content 1",
-                    "document_meta": {
-                        "title": "test title 1",
-                        "source": {"path": "/test/path-1"},
-                        "document_type": "txt",
+                    "id": "1f908deb-bc9f-4b5a-8b73-2e72d8b44dc5",
+                    "text": "test_key 1",
+                    "metadata": {
+                        "content": "test content 1",
+                        "document_meta": {
+                            "title": "test title 1",
+                            "source": {"path": "/test/path-1"},
+                            "document_type": "txt",
+                        },
                     },
                 },
             ),
             models.ScoredPoint(
                 version=1,
                 id="827cad0b-058f-4b85-b8ed-ac741948d502",
-                vector=[0.13, 0.26, 0.30],
+                vector={"text": [0.13, 0.26, 0.30]},
                 score=0.9,
                 payload={
-                    "document": "test_key 2",
-                    "content": "test content 2",
-                    "document_meta": {
-                        "title": "test title 2",
-                        "source": {"path": "/test/path-2"},
-                        "document_type": "txt",
+                    "id": "827cad0b-058f-4b85-b8ed-ac741948d502",
+                    "text": "test_key 2",
+                    "metadata": {
+                        "content": "test content 2",
+                        "document_meta": {
+                            "title": "test title 2",
+                            "source": {"path": "/test/path-2"},
+                            "document_type": "txt",
+                        },
                     },
                 },
             ),
@@ -170,7 +184,6 @@ async def test_list(mock_qdrant_store: QdrantVectorStore) -> None:
     for entry, result in zip(entries, results, strict=True):
         assert entry.metadata["content"] == result["content"]
         assert entry.metadata["document_meta"]["title"] == result["title"]
-        assert entry.vector == result["vector"]
 
 
 def test_create_qdrant_filter() -> None:
@@ -178,9 +191,9 @@ def test_create_qdrant_filter() -> None:
     qdrant_filter = QdrantVectorStore._create_qdrant_filter(where)  # type: ignore
     assert isinstance(qdrant_filter, models.Filter)
     expected_conditions = [
-        models.FieldCondition(key="a", match=models.MatchValue(value="A")),
-        models.FieldCondition(key="b", match=models.MatchValue(value=3)),
-        models.FieldCondition(key="c", match=models.MatchValue(value=True)),
+        models.FieldCondition(key="metadata.a", match=models.MatchValue(value="A")),
+        models.FieldCondition(key="metadata.b", match=models.MatchValue(value=3)),
+        models.FieldCondition(key="metadata.c", match=models.MatchValue(value=True)),
     ]
     assert qdrant_filter.must == expected_conditions
 
@@ -190,8 +203,8 @@ def test_create_qdrant_filter_nested_dict() -> None:
     qdrant_filter = QdrantVectorStore._create_qdrant_filter(where)  # type: ignore
     assert isinstance(qdrant_filter, models.Filter)
     expected_conditions = [
-        models.FieldCondition(key="a", match=models.MatchValue(value="A")),
-        models.FieldCondition(key="b.c", match=models.MatchValue(value="d")),
+        models.FieldCondition(key="metadata.a", match=models.MatchValue(value="A")),
+        models.FieldCondition(key="metadata.b.c", match=models.MatchValue(value="d")),
     ]
     assert qdrant_filter.must == expected_conditions
 
@@ -202,9 +215,9 @@ def test_create_qdrant_filter_with_list() -> None:
     print(qdrant_filter)
     assert isinstance(qdrant_filter, models.Filter)
     expected_conditions = [
-        models.FieldCondition(key="a", match=models.MatchValue(value="A")),
-        models.FieldCondition(key="b[0]", match=models.MatchValue(value="c")),
-        models.FieldCondition(key="b[1]", match=models.MatchValue(value="d")),
+        models.FieldCondition(key="metadata.a", match=models.MatchValue(value="A")),
+        models.FieldCondition(key="metadata.b[0]", match=models.MatchValue(value="c")),
+        models.FieldCondition(key="metadata.b[1]", match=models.MatchValue(value="d")),
     ]
     assert qdrant_filter.must == expected_conditions